# refactor: speed up color