--- a/shpecs/json-redact_shpec.sh
+++ b/shpecs/json-redact_shpec.sh
@@ -6,7 +6,7 @@ describe "json-redact"
   input_file() { echo 'shpecs/support/super_heroes.json'; }
 
   matches_expected 'json-redact --help' <<-EOF
-json-redact (v2023.12.27)
+json-redact (v2024.04.26)
 
 ## json-redact