### Verification

Test files:
 * a random string of `10000` characters
```
cat /dev/urandom | LC_ALL=C tr -dc 'a-z' | head -c 10000 > /tmp/random_chars.txt
```
 * a random list of `1000` 4-letter words to be used as color terms
 ```
 egrep '^[a-z]{4}$' /usr/share/dict/words | shuf -n 1000 > /tmp/random_words.txt
```
<img width="1675" alt="Screenshot 2024-04-22 at 15 09 19" src="https://github.com/zendesk/pravda-client/assets/49626717/ff0160ab-44a4-4a82-85b0-7ab1517c50ed">


#### Baseline tests:
Colorizing our list random of random words in the random chars...

##### First using the current implementation of `color`:
 * `cat /tmp/random_chars.txt | time color $(cat /tmp/random_words.txt)`
 -> `color $(cat /tmp/random_words.txt)  5.48s user 24.69s system 86% cpu 35.047 total`
<img width="1664" alt="Screenshot 2024-04-22 at 14 40 07" src="https://github.com/zendesk/pravda-client/assets/49626717/dd2bfa7d-8795-4882-839c-5811bcbf05fa">

##### Then comparing that to just a single-color `grep`:
 * `cat /tmp/random_chars.txt | time grep -f <(cat /tmp/random_words.txt)`
 -> `grep --color=auto --exclude-dir={.bzr,CVS,.git,.hg,.svn,.idea,.tox} -f   1.65s user 0.00s system 98% cpu 1.682 total`
<img width="1665" alt="Screenshot 2024-04-22 at 14 40 36" src="https://github.com/zendesk/pravda-client/assets/49626717/53f7f98d-824d-4bc3-807c-ddf5fb29347e">

ie. `24.69s` vs `1.65s (user)` - which shows quite an expensive overhead to pay for the pretty colors.

##### Then, after adding the regex color grouping in this PR:
 * `cat /tmp/random_chars.txt | time color $(cat /tmp/random_words.txt)`
  -> `color $(cat /tmp/random_words.txt)  0.28s user 0.40s system 54% cpu 1.252 total`
<img width="1669" alt="Screenshot 2024-04-22 at 14 40 57" src="https://github.com/zendesk/pravda-client/assets/49626717/fc0ec00d-07c0-4885-90f8-42c04fb416c9">

##### And finally avoiding having to count the colors to get the COLOR_COUNT by hard-coding it...
 * `cat /tmp/random_chars.txt | time COLOR_COUNT=25 color $(cat /tmp/random_words.txt)`
    -> `COLOR_COUNT=25 color $(cat /tmp/random_words.txt)  0.24s user 0.22s system 91% cpu 0.512 total`
<img width="1670" alt="Screenshot 2024-04-22 at 14 41 17" src="https://github.com/zendesk/pravda-client/assets/49626717/63ee245a-898f-4f0c-b165-ce014e72575e">

(That looks like just a small saving... ie. from `0.28s (user)` compared to `0.25s (user)` ... so we probably won't bother - but worth leaving the option to do that there anyway so the counted number of colors can be overridden.)

#### Conclusion
 So in this particular test case with `1000` search terms that's a speed up of around 2 orders of magnitude.
 (ie. `25s` vs `0.28s`)
 And all while still keeping the pretty colors!
 <img alt="pretty colors" src="https://i.giphy.com/ruw1bRYN0IXNS.webp">